backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
httpx
pandas
google-generativeai
python-jose[cryptography]
bcrypt
python-multipart
